backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
spacy==3.6.1
transformers==4.35.2
sentence-transformers==2.2.2
supabase==2.0.0
requests==2.31.0
python-dotenv==1.0.0
pathway==0.7.8
numpy==1.24.3
scikit-learn==1.3.2
